requirements.txt
chainer >=7.0.0
funcparserlib >=0.3.6
h5py >=2.6.0
matplotlib >=3.1.1
torch >=1.3.1
pytorch-ignite >=0.2.1
chainer-pytorch-migration >=0.0.1
pytorch-pfn-extras >=0.1.1
